+++ apple-sdk/CHANGELOG.rst
@@ -2,6 +2,11 @@
 apple-sdk Crate Changelog
 =========================
 
+0.3.0
+=====
+
+Not yet released.
+
 0.2.0
 =====